+++ tests/feature/DsmTest.php
@@ -12,7 +12,7 @@ class DsmTest extends \PHPUnit_Framework_TestCase
     public function testCreatesSimpleDsmInHtml()
     {
         $this->assertRegExp(
-            '/PhpDependencies.+X.+8.+1.+11.+29/s',
+            '/PhpDependencies.+X.+8.+1.+1\d.+29/s',
             shell_exec(self::DEPHPEND.' dsm '.self::SRC.' --no-classes -d 2 --format=html')
         );
     }